+++ src/app/components/Header/Auth/Auth.adaptive.css
