+++ src/app/components/Header/AuthForm/AuthForm.adaptive.css
